refactor(analytics): finish shared database usage
