+++ agent/api/app.py
@@ -27,11 +27,11 @@ pin_single_event_loop()
 
 @asynccontextmanager
 async def lifespan(_app: FastAPI) -> AsyncIterator[None]:
+    from agent.analytics.worker import start_worker, stop_worker
+    from agent.dashboard.oauth import validate_github_login_allowlist
     from agent.database.analytics import activate_reporting
     from agent.database.analytics import close as close_analytics
     from agent.database.analytics import migrate as migrate_analytics
-    from agent.analytics.worker import start_worker, stop_worker
-    from agent.dashboard.oauth import validate_github_login_allowlist
     from agent.sandboxes.providers.registry import validate_sandbox_startup_config
     from agent.utils.model import close_cached_models, validate_local_dev_llm_config
